pyproject.toml
[project]
name = "personal-memory-recorder-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "imagehash>=4.3.2",
    "mss>=10.1.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pystray>=0.19.5",
    "pywin32>=311",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "tqdm>=4.67.1",
]
